Added support for the new 1.6 double damage

+++ src/main/java/com/bergerkiller/bukkit/tc/controller/MinecartMember.java
@@ -614,7 +614,7 @@ public abstract class MinecartMember<T extends CommonMinecart<?>> extends Entity
 	}
 
 	@Override
-	public void onDamage(DamageSource damagesource, int damage) {
+	public void onDamage(DamageSource damagesource, double damage) {
 		if (this.entity.isDead()) {
 			return;
 		}
